backend/requirements.txt
Flask
Flask-Dance
requests
python-dotenv 